fix: fix npm start
