+++ src/main.ts
@@ -1,33 +1,34 @@
 import { NestFactory } from '@nestjs/core';
-import helmet from 'helmet';
-import { AppModule } from './app.module';
 import { Express } from "express";
-import { ExpressAdapter } from "@nestjs/platform-express";
 import { INestApplication } from "@nestjs/common";
+import { ExpressAdapter } from "@nestjs/platform-express";
+import helmet from 'helmet';
 
-const port = process.env.PORT || 4000;
+import { AppModule } from './app.module';
 
-async function bootstrap() {
-  const app = await NestFactory.create(AppModule);
-
-  app.enableCors({
-    origin: (req, callback) => callback(null, true),
-  });
-  app.use(helmet());
-
-  await app.listen(port);
-}
-bootstrap().then(() => {
-  console.log('App is running on %s port', port);
-});
+// const port = process.env.PORT || 4000;
+//
+// async function bootstrap() {
+//   const app = await NestFactory.create(AppModule);
+//
+//   app.enableCors({
+//     origin: (req, callback) => callback(null, true),
+//   });
+//   app.use(helmet());
+//
+//   await app.listen(port);
+// }
+// bootstrap().then(() => {
+//   console.log('App is running on %s port', port);
+// });
 
 export async function createApp(
-  expressApp: Express
+    expressApp: Express
 ): Promise<INestApplication> {
   const app = await NestFactory.create(
-    AppModule,
-    new ExpressAdapter(expressApp)
+      AppModule,
+      new ExpressAdapter(expressApp)
   );
+
   return app;
 }
-
